requirements.txt
Django==5.1.5
openai>=2.0.0
python-dotenv>=1.0.0
markdown>=3.5.0
xhtml2pdf>=0.2.13
gunicorn>=21.2.0
psycopg2-binary>=2.9.7
whitenoise>=6.6.0
dj-database-url>=2.1.0
Pillow>=10.0.0